Move last Inspector related classes to the namespace
